Basic Convex Hull Gift Wrapping
- A basic implementation of the gift wrapping algorithm for generating convex hulls was implemented. It still fails for some inputs, it needs to be investigated
- Many API improvements to the MathHelpers functions
- Moved the debug component functions to the module component functions instead of having a separate function
- Added functions to retrieve texture or buffer data from GPU to CPU to be processed here
- Fixed bugs in some functions
- Deleting an entity will wait for any background build function to finish
- Added a new function for GJK to work directly with a point instead of a convex hull

+++ ECSEngine/src/ECSEngine/Utilities/Utilities.h
@@ -4,7 +4,7 @@
 
 namespace ECSEngine {
 
-	ECS_INLINE bool IsPowerOfTwo(int x) {
+	ECS_INLINE bool IsPowerOfTwo(size_t x) {
 		return (x & (x - 1)) == 0;
 	}
 
@@ -254,6 +254,7 @@ namespace ECSEngine {
 	// pointers should be aligned preferably to 32 bytes at least
 	ECSENGINE_API void avx2_copy(void* destination, const void* source, size_t bytes);
 
+	// Vector size needs to be a power of two
 	ECS_INLINE size_t GetSimdCount(size_t count, size_t vector_size) {
 		return count & (-(int64_t)vector_size);
 	}
